third_party/WebKit/LayoutTests/virtual/scroll_customization/fast/scroll-behavior/scroll-customization/scroll-customization-property.html

<!doctype html>
<title> Scroll Customization Property </title>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<style>
* {
  margin: 0;
  padding: 0;
}


p {
  position: relative;
}

#test-div {
  height: 400px;
  width: 400px;
  overflow: scroll;
}
</style>

<div id="test-div">
  <div id="spacer" style="width: 1000px; height:1000px;" ></div>
</div>

<script>
  test(function() {
    assert_true('internals' in window, '"internals" is needed for this test.');
    assert_true('gpuBenchmarking' in chrome,
                '"gpuBenchmarking" is required for this test');
    assert_true(internals.runtimeFlags.scrollCustomizationEnabled,
        '"ScrollCustomization" feature must be enabled.');
    assert_true('ScrollState' in window, "'ScrollState' in window");
  }, 'These tests only work when certain features are enabled.');

  var element = document.getElementById('test-div');
  test(function() {
    assert_true('scrollCustomization' in element.style);
  }, 'Verify "scroll-customization" is part of element\'s style.');

  // Used to track invocation of scroll customization handlers.
  element.didCallApplyScrollHandler = false;
  element.didCallScrollCustomizationHandlers = false;

  var testCases = [
  //  [scroll customization,  gesture,      scroll customization handler called]
      ["pan-up pan-x",        "pan-right",   true],
      ["pan-up pan-x",        "pan-left",    true],
      ["pan-up pan-x",        "pan-up",      true],
      ["pan-up pan-x",        "pan-down",   false],
      ["pan-up pan-left",     "pan-right",  false],
      ["pan-up pan-left",     "pan-left",    true],
      ["pan-up pan-left",     "pan-up",      true],
      ["pan-up pan-left",     "pan-down",   false],
      ["pan-up pan-right",    "pan-right",   true],
      ["pan-up pan-right",    "pan-left",   false],
      ["pan-up pan-right",    "pan-up",      true],
      ["pan-up pan-right",    "pan-down",   false],
      ["pan-up",              "pan-right",  false],
      ["pan-up",              "pan-left",   false],
      ["pan-up",              "pan-up",      true],
      ["pan-up",              "pan-down",   false],
      ["pan-x",               "pan-right",   true],
      ["pan-x",               "pan-left",    true],
      ["pan-x",               "pan-up",     false],
      ["pan-x",               "pan-down",   false],
      ["pan-left",            "pan-right",  false],
      ["pan-left",            "pan-left",    true],
      ["pan-left",            "pan-up",     false],
      ["pan-left",            "pan-down",   false],
      ["pan-right",           "pan-right",   true],
      ["pan-right",           "pan-left",   false],
      ["pan-right",           "pan-up",     false],
      ["pan-right",           "pan-down",   false],
      ["none",                "pan-right",  false],
      ["none",                "pan-left",   false],
      ["none",                "pan-up",     false],
      ["none",                "pan-down",   false],
      ["auto",                "pan-right",   true],
      ["auto",                "pan-left",    true],
      ["auto",                "pan-up",      true],
      ["auto",                "pan-down",    true],
      ["pan-left pan-y",      "pan-right",  false],
      ["pan-left pan-y",      "pan-left",    true],
      ["pan-left pan-y",      "pan-up",      true],
      ["pan-left pan-y",      "pan-down",    true],
      ["pan-right pan-y",     "pan-right",   true],
      ["pan-right pan-y",     "pan-left",   false],
      ["pan-right pan-y",     "pan-up",      true],
      ["pan-right pan-y",     "pan-down",    true],
      ["pan-y",               "pan-right",  false],
      ["pan-y",               "pan-left",   false],
      ["pan-y",               "pan-up",      true],
      ["pan-y",               "pan-down",    true],
      ["pan-down pan-x",      "pan-right",   true],
      ["pan-down pan-x",      "pan-left",    true],
      ["pan-down pan-x",      "pan-up",     false],
      ["pan-down pan-x",      "pan-down",    true],
      ["pan-down pan-left",   "pan-right",  false],
      ["pan-down pan-left",   "pan-left",    true],
      ["pan-down pan-left",   "pan-up",     false],
      ["pan-down pan-left",   "pan-down",    true],
      ["pan-down pan-right",  "pan-right",   true],
      ["pan-down pan-right",  "pan-left",   false],
      ["pan-down pan-right",  "pan-up",     false],
      ["pan-down pan-right",  "pan-down",    true],
      ["pan-down",            "pan-right",  false],
      ["pan-down",            "pan-left",   false],
      ["pan-down",            "pan-up",     false],
      ["pan-down",            "pan-down",    true],
      // The following cases are invalid and should not be set as customization.
      ["pan-left pan-right",  "pan-left",   false],
      ["pan-left pan-right",  "pan-right",  false],
      ["pan-left pan-right",  "pan-up",     false],
      ["pan-left pan-right",  "pan-down",   false],
      ["pan-y pan-y",         "pan-left",   false],
      ["pan-y pan-y",         "pan-right",  false],
      ["pan-y pan-y",         "pan-up",     false],
      ["pan-y pan-y",         "pan-down",   false],
  ];


  function applyScrollCallback(unused_ss) {
    element.didCallApplyScrollHandler = true;
  }

  function distributeScrollCallback(unused_ss) {
    element.didCallDistributeScrollHandler = true;
  }

  element.setApplyScroll(
        applyScrollCallback, 'perform-before-native-scroll');
  element.setDistributeScroll(
        distributeScrollCallback, 'perform-before-native-scroll');


  // Called before each test.
  function resetState() {
    // Make sure the <div> can scroll in all directions. This is necessary or
    // otherwise there won't be a scroll chain containing which includes
    // |element|.
    element.scrollTop =  400;
    element.scrollLeft = 400;

    element.didCallApplyScrollHandler = false;
    element.didCallDistributeScrollHandler = false;
    element.style.scrollCustomization = 'none';
  }

  function didCallScrollCustomizationHandlers() {
    assert_equals(element.didCallApplyScrollHandler, element.didCallDistributeScrollHandler,
                  "It is invalid to call only one of the applyScroll" +
                      " and distributeScroll handler methods.");
    return element.didCallApplyScrollHandler;
  }

  // Returns a promise after applying the |gesture|. The |gesture| will result
  // in a scrolling phase.
  function applyGesture(gesture) {
    var deltaX = 0, deltaY = 0;
    // The delta has to be large enough to avoid getting a fling instead.
    var  delta = 50;
    switch (gesture) {
      case 'pan-right':
      deltaX = -delta;
      break;
      case 'pan-left':
      deltaX = delta;
      break;
      case 'pan-down':
      deltaY = -delta;
      break;
      case 'pan-up':
      deltaY = delta;
      break;
    }

    var startingX = 100, startingY = 100;
    // TODO(ekaramad): The test takes too long due to mandatory delay. Revisit
    // this when and if it becomes possible to create gesture scroll without the
    // delay (https://crbug.com/803168).
    // Necessary to avoid getting a fling instead. Cannot be too small.
    var pauseSeconds = 0.05;
    return new Promise((resolve) => {
      chrome.gpuBenchmarking.pointerActionSequence([{
        source: 'touch',
        actions: [
        { name: 'pointerDown', x: startingX, y: startingY },
        { name: 'pointerMove', x: (startingX + deltaX),
          y: (startingY + deltaY)},
        { name: 'pause', duration: pauseSeconds},
        { name: 'pointerUp'}
        ]
      }], resolve);
    });
  }

  // Define the set of promise tests for all test cases.
  testCases.forEach((testcase) => {
    promise_test(async() => {
      resetState();
      element.style.scrollCustomization = testcase[0];
      await applyGesture(testcase[1]);
      let handlersWereCalled = didCallScrollCustomizationHandlers();
      assert_equals(handlersWereCalled, testcase[2],
        `Handlers were ${handlersWereCalled ? '' : 'not '}called. They must`
        + ` ${testcase[2] ? '' : 'not '}have.`);
    }, `Verify correctness of "scroll-customization: ${testcase[0]}"`
        + ` when gesture "${testcase[1]}" is applied.`);
  });
</script>
